Pacifism finished
Also, fixed the health buff so it actually gets seen as power. (God I hate commas)

With this, the Elf is legally considered "finished", because I CANNOT be bothered to code up Art Appreciator. Now to work on compat with the mods used by the server this was made for!