Get set on fire when you're in the sun
mmm crispy
Can be countered by just not being in the sun, or by wearing anything in the head slot. (also changed the solar weakness to do that last bit.)
Will damage your headgear, though.

(1.17 is untested but I don't think it would break lmao, resources were added in 1.16.2+ so it should be fine)